Move JAX linker to new "link" module closes #188
