+++ theano/link/jax/__init__.py
@@ -1,0 +1,1 @@
+﻿from theano.link.jax.jax_linker import JAXLinker
